Kotlin/Native base commit: b4d5f092a138b9aed2f459e44774c5c0ef18cca7
Kotlin base commit: 926dc085da49018d01aa5cc3cb4b63b7dafd444e

+++ kotlin-native/backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/optimizations/Devirtualization.kt
@@ -820,7 +820,7 @@ internal object Devirtualization {
                 val value = fromId.toLong() or (toId.toLong() shl 32)
                 // This is 64-bit extension of a hashing method from Knuth's "The Art of Computer Programming".
                 // The magic constant is the closest prime to 2^64 * phi, where phi is the golden ratio.
-                val bucketIdx = ((value.toULong() * 11400714819323198393UL) % bagOfEdges.size.toUInt()).toInt()
+                val bucketIdx = ((value.toULong() * 11400714819323198393UL) % bagOfEdges.size.toULong()).toInt()
                 val bucket = bagOfEdges[bucketIdx] ?: LongArrayList().also { bagOfEdges[bucketIdx] = it }
                 for (x in bucket)
                     if (x == value) return
